+++ docs/developers/whats-new.rst
@@ -8,6 +8,6 @@ Unreleased
 
 - Added :py:func:`~pydomcfg.bathymetry.sea_mount`
   useful to generate classic sea mount test case. (:pr:`17`)
-- Added :py:func:`~pydomcfg.utils.generate_nemo_rectilinear_grid`
+- Added :py:func:`~pydomcfg.utils.generate_cartesian_grid`
   useful to generate test datasets. (:pr:`8`)
 - pyDOMCFG is on GitHub! (:pr:`1`)